fix(dropdowns): year dropdown disabled if max date given
All year dropdown options are disabled if a max date is set in
the same year as the max year option. This happens cause the
condition checks from start of the max year means 01.01.xxxx.
If max date is set to e.g. 01.02.2020
»and max year is set to 2020 all options gets disabled.
Therefore changed the condition to check from the end of max year means
31.12.xxx is greater than max date e.g. max date is 01.02.2020
and max year is 2020, the condition would check if 31.12.2020
is after max date.